LayoutTests/fast/text/content-language-case-insensitivity.html

<!DOCTYPE html>
<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta http-equiv="cONtENT-LAnguAGE" content="ja-JP">
<link rel="stylesheet" href="../js/resources/js-test-style.css">
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<p>Test for <a href="https://bugs.webkit.org/show_bug.cgi?id=76701">bug 76701</a>:
map content-language to -webkit-locale. This particular test tests that the
matching of the http-equiv attribute to the content-language state is case
insensitive.
</p>
<div id="console"></div>
<div id="x"></div>
<script>
function languageOfNode(id) {
    var element = document.getElementById(id);
    return window.getComputedStyle(element).webkitLocale;
}
shouldBeEqualToString("languageOfNode('x')", "ja-JP");
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
